feat: Add ability to display environment metrics in Fahrenheit (#1253)
* feat: Add ability to display environment metrics in Fahrenheit

The temperature values in the environment metrics charts and cards are now displayed in Fahrenheit or Celsius based on the user's preference. Celsius is still used as the base unit for calculations and storage.

* Refactor: Rename environmentDisplayTempInFahrenheit to environmentDisplayFahrenheit

Renamed the variable environmentDisplayTempInFahrenheit to environmentDisplayFahrenheit for better clarity and consistency.

* Refactor: Remove unused ENVIRONMENT_METRICS_COLORS

Removed the unused `ENVIRONMENT_METRICS_COLORS` variable from the `EnvironmentMetrics` component.

* Update: Add support for plotting iaq

Added green to the list of colors used for plotting environment metrics to support the newly added iaq readings.